Change in UI and some functionalitites.
Removed all the btnClick() functions, created file - Links.py with class L() that contains function btn1Click() with two parameters. The parameters are  link and file of a software. Each button of the main file, contains two variables: a link, and a file. Variables are passed to the function btn1Click when user clicks.